script to auto-generate schema exports
