+++ mealie/schema/admin/__init__.py
@@ -1,3 +1,4 @@
+# GENERATED CODE - DO NOT MODIFY BY HAND
 from .about import *
 from .backup import *
 from .migration import *
