+++ mealie/schema/cookbook/__init__.py
@@ -1,1 +1,2 @@
+# GENERATED CODE - DO NOT MODIFY BY HAND
 from .cookbook import *
