+++ mealie/schema/events/__init__.py
@@ -1,2 +1,3 @@
+# GENERATED CODE - DO NOT MODIFY BY HAND
 from .event_notifications import *
 from .events import *
