+++ mealie/schema/group/__init__.py
@@ -1,2 +1,10 @@
+# GENERATED CODE - DO NOT MODIFY BY HAND
+from .group import *
+from .group_events import *
+from .group_exports import *
+from .group_migration import *
+from .group_permissions import *
+from .group_preferences import *
 from .group_shopping_list import *
+from .invite_token import *
 from .webhook import *
