+++ mealie/schema/meal_plan/__init__.py
@@ -1,3 +1,4 @@
+# GENERATED CODE - DO NOT MODIFY BY HAND
 from .meal import *
 from .new_meal import *
 from .shopping_list import *
