+++ mealie/schema/recipe/__init__.py
@@ -1,7 +1,15 @@
+# GENERATED CODE - DO NOT MODIFY BY HAND
 from .recipe import *
+from .recipe_asset import *
+from .recipe_bulk_actions import *
 from .recipe_category import *
 from .recipe_comments import *
 from .recipe_image_types import *
 from .recipe_ingredient import *
+from .recipe_notes import *
+from .recipe_nutrition import *
+from .recipe_settings import *
+from .recipe_share_token import *
+from .recipe_step import *
 from .recipe_tool import *
 from .request_helpers import *
