+++ mealie/schema/reports/__init__.py
@@ -1,1 +1,2 @@
+# GENERATED CODE - DO NOT MODIFY BY HAND
 from .reports import *
