+++ mealie/schema/user/__init__.py
@@ -1,3 +1,6 @@
+# GENERATED CODE - DO NOT MODIFY BY HAND
 from .auth import *
+from .registration import *
 from .sign_up import *
 from .user import *
+from .user_passwords import *
